cloud_function/requirements.txt
google-cloud-batch==0.5.0
google-cloud-logging==3.2.5
google-cloud-secret-manager==2.10.0